refactor the clean architecture project
