+++ Blazor.sln
@@ -44,9 +44,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Clean", "Clean", "{3E83C09E
 	ProjectSection(SolutionItems) = preProject
 		Clean\Directory.Build.props = Clean\Directory.Build.props
 		Clean\README.md = Clean\README.md
+		Clean\Clean.slnf = Clean\Clean.slnf
 	EndProjectSection
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BlazorAuthClean.Tests", "Clean\tests\BlazorAuthClean.Tests.csproj", "{E1DF3ECB-1897-4933-BF11-4918A71F04F6}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BlazorAuthClean.Client", "Clean\src\Client\BlazorAuthClean.Client.csproj", "{41E6E281-16C5-4445-8C4F-C6E1ECF47B31}"
 EndProject
@@ -57,6 +56,16 @@ EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BlazorAuthClean.Infrastructure", "Clean\src\Infrastructure\BlazorAuthClean.Infrastructure.csproj", "{2D729F4F-F672-4236-B286-655302366A14}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BlazorAuthClean.Benchmark", "Clean\benchmark\BlazorAuthClean.Benchmark.csproj", "{9340F060-F906-4DCC-8AFC-A875550CF939}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tests", "tests", "{6C61AE3E-E1A4-4757-97D1-6140EEDBFDBD}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BlazorAuthClean.Domain.UnitTests", "Clean\tests\Domain\BlazorAuthClean.Domain.UnitTests.csproj", "{AFCAB499-92EE-4129-8799-368339B4810C}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BlazorAuthClean.Application.IntegrationTests", "Clean\tests\Integration\BlazorAuthClean.Application.IntegrationTests.csproj", "{7D884127-4230-4A04-BE97-FF1595CE7B42}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BlazorAuthClean.Application.UnitTests", "Clean\tests\Application\BlazorAuthClean.Application.UnitTests.csproj", "{BA4B9EED-4425-4BBF-90B7-0BC90EDE97AF}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BlazorAuthClean.Application", "Clean\src\Application\BlazorAuthClean.Application.csproj", "{613EC45E-6F47-4618-94F0-FA109D233ED6}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -112,10 +121,6 @@ Global
 		{F653E92F-A8B0-4124-B8CE-5E8B44A506A2}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{F653E92F-A8B0-4124-B8CE-5E8B44A506A2}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{F653E92F-A8B0-4124-B8CE-5E8B44A506A2}.Release|Any CPU.Build.0 = Release|Any CPU
-		{E1DF3ECB-1897-4933-BF11-4918A71F04F6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E1DF3ECB-1897-4933-BF11-4918A71F04F6}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{E1DF3ECB-1897-4933-BF11-4918A71F04F6}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{E1DF3ECB-1897-4933-BF11-4918A71F04F6}.Release|Any CPU.Build.0 = Release|Any CPU
 		{41E6E281-16C5-4445-8C4F-C6E1ECF47B31}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{41E6E281-16C5-4445-8C4F-C6E1ECF47B31}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{41E6E281-16C5-4445-8C4F-C6E1ECF47B31}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -136,6 +141,22 @@ Global
 		{9340F060-F906-4DCC-8AFC-A875550CF939}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{9340F060-F906-4DCC-8AFC-A875550CF939}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{9340F060-F906-4DCC-8AFC-A875550CF939}.Release|Any CPU.Build.0 = Release|Any CPU
+		{AFCAB499-92EE-4129-8799-368339B4810C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{AFCAB499-92EE-4129-8799-368339B4810C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{AFCAB499-92EE-4129-8799-368339B4810C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{AFCAB499-92EE-4129-8799-368339B4810C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{7D884127-4230-4A04-BE97-FF1595CE7B42}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{7D884127-4230-4A04-BE97-FF1595CE7B42}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{7D884127-4230-4A04-BE97-FF1595CE7B42}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{7D884127-4230-4A04-BE97-FF1595CE7B42}.Release|Any CPU.Build.0 = Release|Any CPU
+		{BA4B9EED-4425-4BBF-90B7-0BC90EDE97AF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{BA4B9EED-4425-4BBF-90B7-0BC90EDE97AF}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{BA4B9EED-4425-4BBF-90B7-0BC90EDE97AF}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{BA4B9EED-4425-4BBF-90B7-0BC90EDE97AF}.Release|Any CPU.Build.0 = Release|Any CPU
+		{613EC45E-6F47-4618-94F0-FA109D233ED6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{613EC45E-6F47-4618-94F0-FA109D233ED6}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{613EC45E-6F47-4618-94F0-FA109D233ED6}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{613EC45E-6F47-4618-94F0-FA109D233ED6}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -153,12 +174,16 @@ Global
 		{A345AB64-A51E-478D-BD6B-4EF1394773AE} = {BC2C2307-7D3A-49D3-952B-2955EF2F78BD}
 		{878D9924-9454-4382-A8C1-6BE7F7D9ACD0} = {BC2C2307-7D3A-49D3-952B-2955EF2F78BD}
 		{F653E92F-A8B0-4124-B8CE-5E8B44A506A2} = {BC2C2307-7D3A-49D3-952B-2955EF2F78BD}
-		{E1DF3ECB-1897-4933-BF11-4918A71F04F6} = {3E83C09E-09EA-4DB1-BF31-E1AD9F9CF318}
 		{41E6E281-16C5-4445-8C4F-C6E1ECF47B31} = {3E83C09E-09EA-4DB1-BF31-E1AD9F9CF318}
 		{94814623-80DA-4632-9A6D-E760DD7AB419} = {3E83C09E-09EA-4DB1-BF31-E1AD9F9CF318}
 		{7F0E254C-9CEF-4408-8D3F-361C7E3C6AA0} = {3E83C09E-09EA-4DB1-BF31-E1AD9F9CF318}
 		{2D729F4F-F672-4236-B286-655302366A14} = {3E83C09E-09EA-4DB1-BF31-E1AD9F9CF318}
 		{9340F060-F906-4DCC-8AFC-A875550CF939} = {3E83C09E-09EA-4DB1-BF31-E1AD9F9CF318}
+		{6C61AE3E-E1A4-4757-97D1-6140EEDBFDBD} = {3E83C09E-09EA-4DB1-BF31-E1AD9F9CF318}
+		{AFCAB499-92EE-4129-8799-368339B4810C} = {6C61AE3E-E1A4-4757-97D1-6140EEDBFDBD}
+		{7D884127-4230-4A04-BE97-FF1595CE7B42} = {6C61AE3E-E1A4-4757-97D1-6140EEDBFDBD}
+		{BA4B9EED-4425-4BBF-90B7-0BC90EDE97AF} = {6C61AE3E-E1A4-4757-97D1-6140EEDBFDBD}
+		{613EC45E-6F47-4618-94F0-FA109D233ED6} = {3E83C09E-09EA-4DB1-BF31-E1AD9F9CF318}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {6BA7374B-6AB0-4CD0-A1CB-13A60C9641D6}
