+++ Clean/src/Client/BlazorAuthClean.Client.csproj
@@ -14,7 +14,7 @@
 	</ItemGroup>
 
 	<ItemGroup>
-	  <ProjectReference Include="..\Domain\BlazorAuthClean.Domain.csproj" />
+		<ProjectReference Include="..\Domain\BlazorAuthClean.Domain.csproj"/>
 	</ItemGroup>
 
 </Project>
