+++ Clean/src/Domain/Identity/ApplicationUser.cs
