+++ Clean/src/Infrastructure/Persistence/ApplicationDbContext.cs
@@ -1,7 +1,9 @@
+using BlazorAuthClean.Data;
+
 using Microsoft.AspNetCore.Identity.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore;
 
-namespace BlazorAuthClean.Data
+namespace BlazorAuthClean.Infrastructure.Persistence
 {
 	public class ApplicationDbContext(DbContextOptions<ApplicationDbContext> options) : IdentityDbContext<ApplicationUser>(options)
 	{
