+++ Clean/tests/Domain/BlazorAuthClean.Domain.UnitTests.csproj
@@ -10,10 +10,7 @@
 	</PropertyGroup>
 
 	<ItemGroup>
-		<PackageReference Include="Microsoft.NET.Test.Sdk" />
-		<PackageReference Include="xunit" />
-		<PackageReference Include="xunit.runner.visualstudio" PrivateAssets="all"/>
-		<PackageReference Include="coverlet.collector" PrivateAssets="all"/>
+		<ProjectReference Include="..\..\src\Domain\BlazorAuthClean.Domain.csproj"/>
 	</ItemGroup>
 
 </Project>
