+++ Directory.Build.props
@@ -17,5 +17,16 @@
         <!--        <ManagePackageVersionsCentrally>true</ManagePackageVersionsCentrally>-->
     </PropertyGroup>
 
+	<!-- xUnit Tests-->
+	<ItemGroup Condition="$(MSBuildProjectName.Contains('Tests'))">
+		<PackageReference Include="Microsoft.AspNetCore.TestHost" />
+		<PackageReference Include="Microsoft.NET.Test.Sdk" />
+		<PackageReference Include="Moq" />
+		<PackageReference Include="xunit" />
+		<PackageReference Include="xunit.runner.visualstudio" />
+		<PackageReference Include="coverlet.collector" />
+		<PackageReference Include="coverlet.msbuild" />
+	</ItemGroup>
+
 
 </Project>
