+++ Directory.Build.targets
@@ -21,10 +21,13 @@
         <PackageReference Update="Microsoft.EntityFrameworkCore.Tools" Version="$(FrameworkVersion)"/>
 
 		<!--tests-->
+		<PackageReference Update="Microsoft.AspNetCore.TestHost" Version="$(FrameworkVersion)"/>
 		<PackageReference Update="Microsoft.NET.Test.Sdk" Version="17.7.2"/>
+		<PackageReference Update="Moq" Version="4.20.69"/>
 		<PackageReference Update="xunit" Version="2.5.3"/>
 		<PackageReference Update="xunit.runner.visualstudio" Version="2.5.3" PrivateAssets="all"/>
 		<PackageReference Update="coverlet.collector" Version="6.0.0" PrivateAssets="all"/>
+		<PackageReference Update="coverlet.msbuild" Version="6.0.0" PrivateAssets="all"/>
 
 		<!--benchmark-->
 		<PackageReference Update="BenchmarkDotNet" Version="0.13.9"/>
